Move block formatting options into block menu
