+++ editor/plugins/withNormalization.ts
@@ -1,6 +1,7 @@
 import { Editor, Element, Node, Selection, Text, Transforms } from 'slate';
 import { ElementType, Mark } from 'types/slate';
 import { isListType } from 'editor/formatting';
+import { isTextType } from 'editor/checks';
 
 const withNormalization = (editor: Editor) => {
   return withListNormalization(withInlineNormalization(editor));
@@ -66,11 +67,7 @@ const withListNormalization = (editor: Editor) => {
       for (const [child, childPath] of Node.children(editor, path)) {
         if (
           Element.isElement(child) &&
-          (child.type === ElementType.Paragraph ||
-            child.type === ElementType.HeadingOne ||
-            child.type === ElementType.HeadingTwo ||
-            child.type === ElementType.HeadingThree ||
-            child.type === ElementType.ListItem)
+          (isTextType(child.type) || child.type === ElementType.ListItem)
         ) {
           Transforms.unwrapNodes(editor, { at: childPath });
           return;
@@ -81,13 +78,7 @@ const withListNormalization = (editor: Editor) => {
     // Convert paragraphs and headings to list items if they are the children of a list
     if (Element.isElement(node) && isListType(node.type)) {
       for (const [child, childPath] of Node.children(editor, path)) {
-        if (
-          Element.isElement(child) &&
-          (child.type === ElementType.Paragraph ||
-            child.type === ElementType.HeadingOne ||
-            child.type === ElementType.HeadingTwo ||
-            child.type === ElementType.HeadingThree)
-        ) {
+        if (Element.isElement(child) && isTextType(child.type)) {
           Transforms.setNodes(
             editor,
             { type: ElementType.ListItem },
